refactor: address branding PR review
- validate theme.json with zod instead of an unchecked cast
- extract a shared GameBackground component + imageFallback helper
- move branding/branding.ts to src/branding.ts
- drop the redundant branding.example folder
- trim comments and fix lint/format on vite.config.ts
