--- branding.example/theme.json
@@ -1,15 +1,0 @@
-{
-  "appName": "My Quiz",
-  "colors": {
-    "primary": "#ff9900",
-    "secondary": "#1a140b"
-  },
-  "answerColors": ["#e69f00", "#56b4e9", "#3dbfa0", "#cc79a7"],
-  "font": {
-    "family": "Rubik",
-    "url": "https://fonts.googleapis.com/css2?family=Rubik:wght@300..900&display=swap"
-  },
-  "logo": "/branding/logo.svg",
-  "favicon": "/branding/favicon.svg",
-  "background": "/branding/background.png"
-}
