+++ packages/web/src/main.tsx
@@ -1,5 +1,5 @@
 import "@fontsource-variable/rubik/wght.css"
-import { applyBranding, loadBranding } from "@razzia/web/branding/branding"
+import { applyBranding, loadBranding } from "@razzia/web/branding"
 import Toaster from "@razzia/web/components/Toaster"
 import { socketClient } from "@razzia/web/features/game/contexts/socket-context"
 import "@razzia/web/i18n"
